Commit for #376 - much more to come

+++ src/PolyGlot/IOHandler.java
@@ -934,7 +934,7 @@ public class IOHandler {
         try (InputStream tmp = this.getClass().getResourceAsStream(PGTUtil.ButtonFontLocation)) {
             GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment();
             ret = Font.createFont(Font.TRUETYPE_FONT, tmp);
-            ret = ret.deriveFont(new Float(14)); // default to size 12 font
+            ret = ret.deriveFont(new Float(12)); // default to size 12 font
             if (ret != null) {
                 ge.registerFont(ret);
             }
